Fix up copy paste actions and reducer
